+++ backend/canisters/user_index/api/src/queries/external_achievements.rs
@@ -26,7 +26,7 @@ pub struct SuccessResult {
 #[ts_export(user_index, external_achievements)]
 #[derive(CandidType, Debug)]
 pub struct ExternalAchievement {
-    pub id: u128,
+    pub id: u32,
     pub name: String,
     pub url: String,
     pub chit_reward: u32,
